+++ pkg/streamingpromql/query.go
@@ -66,7 +66,7 @@ type Query struct {
 	result *promql.Result
 }
 
-func (e *Engine) newQuery(ctx context.Context, queryable storage.Queryable, opts promql.QueryOpts, timeRange types.QueryTimeRange) (*Query, error) {
+func (e *Engine) newQuery(ctx context.Context, queryable storage.Queryable, opts promql.QueryOpts, timeRange types.QueryTimeRange, originalExpression string) (*Query, error) {
 	if opts == nil {
 		opts = promql.NewPrometheusQueryOpts(false, 0)
 	}
@@ -81,7 +81,7 @@ func (e *Engine) newQuery(ctx context.Context, queryable storage.Queryable, opts
 		return nil, fmt.Errorf("could not get memory consumption limit for query: %w", err)
 	}
 
-	memoryConsumptionTracker := limiter.NewMemoryConsumptionTracker(maxEstimatedMemoryConsumptionPerQuery, e.queriesRejectedDueToPeakMemoryConsumption)
+	memoryConsumptionTracker := limiter.NewMemoryConsumptionTracker(maxEstimatedMemoryConsumptionPerQuery, e.queriesRejectedDueToPeakMemoryConsumption, originalExpression)
 	stats, err := types.NewQueryStats(timeRange, e.enablePerStepStats && opts.EnablePerStepStats(), memoryConsumptionTracker)
 	if err != nil {
 		return nil, err
@@ -94,6 +94,7 @@ func (e *Engine) newQuery(ctx context.Context, queryable storage.Queryable, opts
 		stats:                    stats,
 		topLevelQueryTimeRange:   timeRange,
 		lookbackDelta:            lookbackDelta,
+		originalExpression:       originalExpression,
 	}
 
 	return q, nil
@@ -121,12 +122,11 @@ func (e *Engine) newQueryFromExpression(ctx context.Context, queryable storage.Q
 		}
 	}
 
-	q, err := e.newQuery(ctx, queryable, opts, timeRange)
+	q, err := e.newQuery(ctx, queryable, opts, timeRange, qs)
 	if err != nil {
 		return nil, err
 	}
 
-	q.originalExpression = qs
 	q.statement = &parser.EvalStmt{
 		Expr:          expr,
 		Start:         start,